Initial commit
Changed the names of the css and html files. Added the div elements inside of the game box div. Added the ids of butterfly, bat, fox.  Added the transition css style for the navigation links. Added the linear gradient for the navigation links. Made the links opacity of 1 .  Made each div element with the id of bat butterfly, fox have a background of the image of the animal, be left center have no titling, and have the image cover the page.