Added animation styles/ fixed background issue
Added the playSprite, flyRight animations to the butterfly div. Added the playSprite, flyRight animations to the bat div. Added the playSprite animation to the fox div. Changed the keyframe plaSprite animation from background to background-position. Added the keyframe flyRight for 25% top has 150px, 50% top to 55px, 65% top to 120px;, 90% top to 50px, 100% top is 80px, left coordinate to 100%.